utf8_heavy: Return pointer to special treatment hash.
Many of the mapping tables that utf8_heavy.pl deals with have maps that
don't fit into the general scheme of things, and a hash is generated for
those.  This patch causes utf8_heavy to add an entry to the object it
generates containing a reference to this hash.  Prior to this, the name
of the hash had to be known to the users of the table.