mktables: Put off removing the \N{BELL} conflict
Unicode 6.0 re-used the name BELL for a character other than what it
was traditionally for.  There is code in mktables to change the
treatment for 5.15, so that it wouldn't be forgotten.  But I'm thinking
that we might want to wait til 5.18 to change, and so I just changed it
to 5.17 instead.