release_managers_guide: replace =item with =head3
The main body of the Release Manager's Guide is currently just a long list
of steps, each with a '=item *'. Replace these with
    =head3 some description
so that its easier to navigate around.

Also do a few minor structural fixups along the way (eg splitting or
merging individual =item sections).

Also a few typos and bits fixed as I noticed them.