Use L<> rather than C<> in the Updated Modules section of perldelta.
Previous perldeltas have been inconsistent with each other as to whether to
use L<> to C<> (but internally consistent). perldelta5140.pod uses L<>, which
is more useful, as it provides a direct link to fuller documentation at the
point where the reader is likely to want to follow it.

Switch the two existing entries from C<> to L<>, and update the perldelta
template to suggest L<>.