perlhack.pod: invoke git-format-patch with --attach
As George Greer noted on p5p, --attach causes the message to be
written using MIME-attach syntax, so when perlbug sends it,
rt.perl.org detaches the file and adds it to the ticket.

While tradtional inline patches appear to survive without whitespace
mangling, attachments are more in keeping with RTs design and use.

Signed-off-by: Jim Cromie <jim.cromie@gmail.com>