eintr.t: skip based on capability rather than OS
The t/io/eintr.t tests require read/write system to calls to be
interruptible (to see if anything nasty can be done by the signal
handler).

Many platforms aren't interruptible, which means the tests would hang.
We currently work round this by skipping based on a hard-coded list of
OSes (such as win32, VMS etc).

Change this so that we instead do an initial test as to whether they are
interruptible, and if not, skip the whole test file.