perl_clone_using: do simple assign/copying first
For all the elements of the interpreter that can be cloned by
a simple initialisation (e.g. assign NULL/0/-1), or by a direct
copy from proto, do these first. This means that anything more complex,
that might for example require a malloc, gets done after as much of the
interpreter structure has already been initialised as possible.

This is the beginnings of a fix for [perl #90006].

This commit should have no change in functionality; it's simply moving a
bunch of assign statements to earlier within perl_clone_using().