perldelta: Add potential entries
I looked through all my commits, and came up with this text for all the
ones I think could possibly be desired to put into the perldelta.  All
other commits I've made for 5.15 so far, can be ignored IMO as far as
the delta is concerned.  I'm not so sure about even the ones I've
done here.  Feel free to omit them.  The wording for the edge case is
convoluted; but it's hard for me to describe the complicated
circumstances when the bug appeared.