t/op/alarm.t: reduce intermittent failures
Fix a race condition where the alarm is reset *after* the handler
and an eval have been removed.

Also, make the points at which time is measured closer to the things
they need to measure, to reduce timing noise on a heavily-loaded system.

Hopefully these measures will reduce the number of random smoke failures,
although I couldn't reproduce the failures, even on a heavily-loaded
machine.